requirements.txt
alpaca-py
pandas
pandas_ta
yfinance
requests
pytz